Move infusions to dedicated dir
