+++ src/mix/data-admix.spec.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from '@jest/globals';
 import { DataFaucet } from '../data-faucet.js';
-import { withValue } from '../with-value.js';
+import { withValue } from '../infusions/with-value.js';
 import { admix } from './data-admix.js';
 import { DataMixer } from './data-mixer.js';
