+++ src/mod.ts
@@ -1,7 +1,6 @@
 export * from './data-faucet.js';
 export * from './data-infusion.js';
 export * from './data-sink.js';
+export * from './infusions/mod.js';
 export * from './joints/mod.js';
 export * from './mix/mod.js';
-export * from './with-all.js';
-export * from './with-value.js';
